test(gda-balancing): move the subprocess tier to test_e2e_cli.py (#504)
Align with the family's e2e naming convention (gda's test_e2e_*): the
true-subprocess tests (entry-point agreement, stream separation, key
user path) get their own module. Pure move — no behavior change; this
tier stays unmarked and runs in the standard CI job (no engine
dependency, unlike gda's gated Godot e2e).
